ex2/requirements.txt
africastalking==1.2.4
altgraph==0.17.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
contextlib2==21.6.0
cryptography==39.0.0
engineering-notation==0.8.0
evdev==1.6.1
idna==3.4
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.4.0
pyAesCrypt==6.0.0
pycparser==2.21
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pynput==1.7.6
pysteganography==0.1.0
python-xlib==0.33
requests==2.28.2
schema==0.7.5
six==1.16.0
steganography==0.1.1
tk-tools==0.16.0
urllib3==1.26.14
google-api-python-client